Fix channel space types and walletconnect alias
