+++ src/fidgets/farcaster/Feed.tsx
@@ -23,7 +23,6 @@ import {
   type FidgetSettingsStyle,
 } from "@/common/fidgets";
 import useLifoQueue from "@/common/lib/hooks/useLifoQueue";
-import { FeedType } from "@neynar/nodejs-sdk/build/api";
 import { isNil } from "lodash";
 import { BsChatRightHeart, BsChatRightHeartFill } from "react-icons/bs";
 import { useInView } from "react-intersection-observer";
@@ -31,12 +30,7 @@ import { useFarcasterSigner } from ".";
 import { mobileStyleSettings, WithMargin } from "../helpers";
 import { CastRow } from "./components/CastRow";
 import { CastThreadView } from "./components/CastThreadView";
-
-export enum FilterType {
-  Channel = "channel_id",
-  Users = "fids",
-  Keyword = "keyword",
-}
+import { FEED_TYPES, FilterType, FeedType } from "./feedConstants";
 
 export type FeedFidgetSettings = {
   feedType: FeedType;
@@ -52,11 +46,6 @@ export type FeedFidgetSettings = {
   membersOnly?: boolean;
 } & FidgetSettingsStyle;
 
-const FILTER_TYPES = [
-  { name: "Channel", value: FilterType.Channel },
-  { name: "Users", value: FilterType.Users },
-  { name: "Keyword", value: FilterType.Keyword },
-];
 
 export const FilterTypeSelector: React.FC<{
   onChange: (value: string) => void;
@@ -327,15 +316,6 @@ const feedProperties: FidgetProperties<FeedFidgetSettings> = {
   mobileIcon: <BsChatRightHeart size={20} />,
   mobileIconSelected: <BsChatRightHeartFill size={20} />,
 };
-
-export const FEED_TYPES = [
-  { name: "Following", value: FeedType.Following },
-  { name: "For you", value: "for_you" },
-  { name: "Trending", value: "trending" },
-  { name: "Filter", value: FeedType.Filter },
-];
-
-export { FeedType };
 
 const Feed: React.FC<FidgetArgs<FeedFidgetSettings>> = ({ settings }) => {
   const {
